requirements.txt.txt
tensorflow
matplotlib
Pillow
opendatasets
numpy
scikit-learn